Restructure directories and don't dynamically include CSS.
Now all example code is under the 'example' directory.

The CSS included a relative URL to webmaker-logo.png, which didn't
work when dynamically injecting the stylesheet in the requirejs module. So
now the CSS isn't dynamically injected, but this requires the client to
manually include a <link> tag to the stylesheet in their HTML.