Display of logged-in/out specific content is controlled by CSS now.
This is done by toggling a 'logged-in' class on the nav container and
defining CSS rules to control the display of the persona login and user menu,
which relieves JS of having to control them directly. This will also make
adding more login-sensitive content, like the badge widget, easier to add
in the future.