Added pointy end of popover.
This is taken from @k88hudson's ridiculously crafty 'tooptip' code in
webmaker-nav.css and slightly modified. Ultimately we might want to
refactor things to reduce duplication of the CSS. (The only thing
really modified from tooptip's pointer is setting the 'left' CSS property
so it's on the left end of the top edge of the popover rather than
centered.)

Also, I have no idea how the pointy thing actually works. I sat there
staring at the CSS code for a while, but I could not figure it out, so I
just ended up copy-pasting the code and tweaking things until they looked
ok. As a result, this solution might be suboptimal in ways that I'm
completely unaware of.