Clicking '+ Backpack' now pushes to the backpack via the issuer API.
I've structured the code to be as fault-tolerant of OpenBadges failing
as possible, as well as not blocking on loading of the issuer JS API. So
right now the "+ Backpack" button simply won't be displayed if
window.OpenBadges is falsy, which decouples loading of the issuer API
from the loading of core application logic. It also makes it possible
for an embedder to opt-out of backpack integration if needed (though we
don't expect this to ever be the case).